chore(clippy): resolve all warnings under -D warnings
- Replace runtime-constant assertions with `const _: () = assert!(...)`
  in con_002 / con_003, keeping a runtime smoke test backed by non-const
  locals.
- Drop `clone()` calls on Copy types (EpochPhase, DfspCloseSnapshot,
  PublicKey) in tests.
- Remove redundant `as u64` casts where the lhs is already u64
  (rew_001 / rew_005 / end_to_end_test).

`cargo clippy --all-targets -- -D warnings` is clean; `cargo fmt --check`
is clean; `cargo test` reports 63 green suites / 470+ passing tests.

+++ tests/integration/end_to_end_test.rs
@@ -104,7 +104,7 @@ fn test_full_epoch_lifecycle_via_public_api() {
     assert!(m.should_advance(133));
 
     // Reward distribution derived from protocol rules.
-    let base_reward: u64 = (1..=(BLOCKS_PER_EPOCH as u64))
+    let base_reward: u64 = (1..=BLOCKS_PER_EPOCH)
         .map(|h| total_block_reward(h, h == 1))
         .sum();
     let total_reward = epoch_reward_with_floor(base_reward);
